+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## [Unreleased]
+## [v0.5.0]
+
+### Added
+- Type and Validation Support from the library `@the-devoyage/request-filter-language`
+
+### Removed
+- GraphQL Types - You may now get the GraphQL Schema from `@the-devoyage/request-filter-language`.
+- Validators - You may now import Validation Schemas from `@the-devoyage/request-filter-language`.
+- Parse Field Filters - Now imported from `@the-devoyage/request-filter-language`.
+- Typescript Types Associated with Field Filters - Now import from `@the-devoyage/request-filter-language`.
 
 ## [v0.4.0]